Replace Select with native to fix e2e test
