+++ app/e2e/pro.spec.ts
@@ -1,112 +1,22 @@
-import { expect, Page, test } from "@playwright/test";
-import jsdom from "jsdom";
+import { test, expect } from "@playwright/test";
+import { BASE_URL, TESTING_EMAIL_PRO, TESTING_PASS_PRO } from "./utils";
 import path from "path";
 
-import { BASE_URL, getTempEmail, getTempEmailMessage } from "./utils";
-
-let page: Page;
-let email = "";
-let publicLinkValue: string | null = "";
-
-test.describe.configure({
-  mode: "serial",
-});
-
-test.skip(({ browserName }) => browserName !== "chromium", "Chromium only!");
-
-test.beforeAll(async ({ browser }) => {
-  page = await browser.newPage();
+/* Log In */
+test.beforeEach(async ({ page }) => {
   await page.goto(BASE_URL);
+  await page.getByRole("link", { name: "Log In" }).click();
+  await page.getByTestId("sign-in-email").click();
+  await page.getByTestId("sign-in-email").fill(TESTING_EMAIL_PRO);
+  await page.getByTestId("sign-in-email").press("Tab");
+  await page.getByTestId("sign-in-password").fill(TESTING_PASS_PRO);
+  await page.getByTestId("sign-in-email-pass").click();
+  await expect(page.getByRole("link", { name: "Account" })).toBeVisible();
+  // Wait for the Upgrade link to dissappear because it means we know the user is PRO
+  await page.getByTestId("pro-link").waitFor({ state: "detached" });
 });
 
-test.afterAll(async () => {
-  await page.close();
-});
-
-test("Sign Up", async () => {
-  // set timeout
-  test.setTimeout(240000);
-
-  await page.getByRole("link", { name: "Flowchart Fun Pro" }).click();
-  await expect(page).toHaveURL(`${BASE_URL}/pricing`);
-
-  await page.getByTestId("yearly-plan-button").click();
-  await page.getByRole("button", { name: "Continue" }).click();
-
-  email = await getTempEmail();
-  await page.getByTestId("email-input").fill(email);
-  await page.getByRole("button", { name: "Continue" }).click();
-
-  const iframe = page.frameLocator(
-    'internal:attr=[title="Secure payment input frame"i]'
-  );
-
-  await iframe
-    .getByPlaceholder("1234 1234 1234 1234")
-    .fill("4242 4242 4242 4242");
-  await iframe.getByPlaceholder("MM / YY").fill("05 / 50");
-
-  await iframe.getByPlaceholder("CVC").fill("222");
-  await iframe.getByRole("combobox", { name: "Country" }).selectOption("US");
-  await iframe.getByPlaceholder("12345").fill("12345");
-
-  await page.getByRole("button", { name: "Sign Up" }).click();
-
-  // expect the url to start with "/l"
-  await expect(page).toHaveURL(new RegExp(`${BASE_URL}/l`));
-
-  // expect sign up success to be on the screen
-  // make sure test id welcome-message is visible
-  await expect(page.getByTestId("welcome-message")).toBeVisible();
-
-  /* Part 2: Get Auth Email */
-  // fill in testid sign-in-magic-email with email
-  await page.getByTestId("sign-in-magic-email").fill(email);
-
-  // get button with test id "request-magic-link"
-  await page.getByTestId("request-magic-link").click();
-  await expect(
-    page.getByText(/Check your email for a link to log in/i)
-  ).toBeVisible({ timeout: 60000 });
-
-  // wait for email
-  let emails = [],
-    i = 0;
-  while (emails.length === 0 && i < 20) {
-    // Wait 5 seconds
-    await new Promise((resolve) => setTimeout(resolve, 10000));
-    // Check for email
-    const response = await getTempEmailMessage(email);
-    if (!("error" in response)) {
-      emails = response;
-    }
-    i++;
-  }
-
-  expect(emails.length).toBeGreaterThanOrEqual(1);
-  const signInEmailIndex = emails.findIndex((email: { mail_html: string }) =>
-    /supabase.co\/auth\/v1\/verify/.test(email.mail_html)
-  );
-
-  expect(signInEmailIndex).toBeGreaterThanOrEqual(0);
-
-  // apply this to a new document to get link and then load link and confirm the Account link present
-  const html = emails[signInEmailIndex].mail_html;
-  const { window } = new jsdom.JSDOM(html);
-  const { document } = window;
-  const link = document.querySelector("a");
-  expect(link).toBeTruthy();
-  expect(link?.href).toBeTruthy();
-  await page.goto(link?.href as string);
-
-  // reload page
-  await page.reload();
-
-  // expect link with "Account" to be present
-  await expect(page.getByText("Account")).toBeVisible({ timeout: 60 * 1000 });
-});
-
-test("Create new chart", async () => {
+test("Create new chart", async ({ page }) => {
   await page.getByRole("link", { name: "New" }).click();
   await page.getByLabel("Name Chart").fill("my new chart");
   await page.getByRole("button", { name: "Create" }).click();
@@ -114,14 +24,15 @@ test("Create new chart", async () => {
   await expect(page).toHaveURL(new RegExp(`${BASE_URL}/u/\\d+`));
 });
 
-test("Rename Chart", async () => {
+test("Rename Chart", async ({ page }) => {
   await page.getByLabel("Rename").click();
   await page.getByRole("textbox").fill("to publish");
   await page.getByRole("button", { name: "Rename" }).click();
   await expect(page.getByLabel("Rename")).toHaveText("to publish");
 });
 
-test("Publish Chart", async () => {
+test("Publish Chart & Clone from Public", async ({ page }) => {
+  let publicLinkValue: string | null = "";
   await page.getByRole("button", { name: "Export" }).click();
   await page.getByLabel("Make publicly accessible").check();
 
@@ -133,9 +44,7 @@ test("Publish Chart", async () => {
   publicLinkValue = await publicLink.getAttribute("value");
 
   expect(publicLinkValue).toBeTruthy();
-});
 
-test("Clone from Public Link", async () => {
   if (!publicLinkValue) throw new Error("publicLinkValue is not set");
 
   // navigate to public url
@@ -156,13 +65,13 @@ test("Clone from Public Link", async () => {
   await page1.close();
 });
 
-test("Open Chart From Charts Page", async () => {
+test("Open Chart From Charts Page", async ({ page }) => {
   await page.goto(BASE_URL);
   await page.getByRole("link", { name: "Charts" }).click();
   await page.getByRole("link", { name: /to publish.*/gi }).click();
 });
 
-test("Download SVG", async () => {
+test("Download SVG", async ({ page }) => {
   await page.getByLabel("Export").click();
 
   // Click [aria-label="Download SVG"]
@@ -173,13 +82,11 @@ test("Download SVG", async () => {
   await page.getByTestId("close-button").click();
 });
 
-test("Go to Sandbox. Save Sandbox Chart", async () => {
+test("Go to Sandbox. Save Sandbox Chart", async ({ page }) => {
   await page.goto(BASE_URL);
 
-  // Wait for the upgrade to pro to disappear, so we know customer has loaded
-  await page
-    .getByRole("link", { name: "Upgrade to Pro" })
-    .waitFor({ state: "detached" });
+  // wait for the test-id "pro-link" to disappear
+  await page.getByTestId("pro-link").waitFor({ state: "detached" });
 
   await page.getByRole("button", { name: "Save" }).click();
   await page.getByLabel("Title").click();
@@ -190,7 +97,7 @@ test("Go to Sandbox. Save Sandbox Chart", async () => {
   await expect(page).toHaveURL(new RegExp(`${BASE_URL}/u/\\d+`));
 });
 
-test("Create chart from ai", async () => {
+test("Create chart from ai", async ({ page }) => {
   await page.getByRole("link", { name: "New" }).click();
   await page.getByTestId("Use AI").click();
   await page.locator('textarea[name="subject"]').click();
@@ -204,7 +111,7 @@ test("Create chart from ai", async () => {
   });
 });
 
-test("Create chart from imported data", async () => {
+test("Create chart from imported data", async ({ page }) => {
   try {
     await page.getByRole("link", { name: "New" }).click();
     await page.getByRole("button", { name: "Create" }).click();
@@ -230,25 +137,18 @@ test("Create chart from imported data", async () => {
     // await page.getByTestId("import-data-file-uploader").click();
     await page.getByTestId("import-data-file-uploader").setInputFiles(filePath);
 
-    await page.getByTestId("node-label-select").press("Enter");
     await page
-      .getByRole("option", { name: "Process Step Description" })
-      .press("Enter");
-
+      .getByTestId("node-label-select")
+      .selectOption("Process Step Description");
     await page.getByTestId("edges-in-source-node-row").click();
-
-    await page.getByTestId("target-column-select").press("Enter");
-    await page.getByRole("option", { name: "Next Step ID" }).press("Enter");
-
+    await page.getByTestId("target-column-select").selectOption("Next Step ID");
+    await page.getByTestId("target-delimiter-input").click();
     await page.getByTestId("target-delimiter-input").fill(",");
+    await page
+      .getByTestId("edge-label-column-select")
+      .selectOption("Connector Label");
+    await page.getByTestId("import-submit-button").click();
 
-    await page.getByTestId("edge-label-column-select").press("Enter");
-    await page.getByRole("option", { name: "Connector Label" }).press("Enter");
-
-    // click test id "import-submit-button"
-    await page.getByTestId("import-submit-button").click({ force: true });
-
-    // expect "You are about to add 9 nodes and 10 edges to your graph." to be visible
     await expect(
       page.getByText("You are about to add 9 nodes and 10 edges to your graph.")
     ).toBeVisible();
