+++ app/e2e/utils.ts
@@ -89,4 +89,10 @@ if (!TESTING_EMAIL || !TESTING_PASSWORD) {
   throw new Error("Must set TESTING_EMAIL and TESTING_PASS");
 }
 
-export { TESTING_EMAIL, TESTING_PASSWORD };
+const TESTING_EMAIL_PRO = process.env.TESTING_EMAIL_PRO as string;
+const TESTING_PASS_PRO = process.env.TESTING_PASS_PRO as string;
+if (!TESTING_EMAIL_PRO || !TESTING_PASS_PRO) {
+  throw new Error("Must set TESTING_EMAIL_PRO and TESTING_PASS_PRO");
+}
+
+export { TESTING_EMAIL, TESTING_PASSWORD, TESTING_EMAIL_PRO, TESTING_PASS_PRO };
